pyproject.toml
[project]
name = "restaurant-voice-agent"
version = "0.1.0"
description = "LiveKit multiagent restaurant voice assistant"
requires-python = ">=3.11"
dependencies = [
    # LiveKit — single package with extras, current stable is 1.4.x
    "livekit-agents[openai,deepgram,cartesia,silero,turn-detector]~=1.4",

    # Config & validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "logfire[pydantic]>=4.29.0",


    # YAML prompt loading
    "pyyaml>=6.0.0",

    # Structured logging
    "structlog>=24.0.0",

    # Utils
    "httpx>=0.27.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ✅ ADD THIS — tells hatchling where your src package lives
[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"